requirements.txt
# Vex RAG Plugin Dependencies
# Python 3.11+

# Vector Database
lancedb>=0.5.0

# Embeddings & ML Models
sentence-transformers>=2.2.2
torch>=2.0.0

# Ollama Integration
ollama>=0.1.0

# MCP Server
fastmcp>=0.3.0

# Configuration & Environment
pyyaml>=6.0
python-dotenv>=1.0.0

# PII Sanitization
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.7.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.0.0
python-pptx>=0.6.0

# Search & Ranking
rank-bm25>=0.2.2

# Development Dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
ruff>=0.0.285
